// emasaku-cloudflare--semua-menjadi-satu/types.ts
export enum UserRole {
  MEMBER = 'Member',
  ADMIN = 'Admin',
  DEVELOPER = 'Developer',
}

export interface User {
  id: string;
  name: string;
  email?: string;
  avatarUrl: string;
  role: UserRole;
  joinedDate: string;
  whatsapp?: string;
  driverBalance?: number;
  isRetailPartner?: boolean;
  isLaundryPartner?: boolean;
  isSeller?: boolean;
  isWastePartner?: boolean;
}
// ...
export enum LaundryServiceType {
    KILOAN = 'Kiloan',
    SATUAN = 'Satuan',
    EKSPRES = 'Ekspres'
}

export interface LaundryOrderItem {
    name: string;
    quantity: number;
}

export type LaundryStatus = 
    'Pesanan Dibuat' | 
    'Mencari Kurir' | 
    'Kurir Menuju Lokasi' |
    'Pakaian Diterima' |
    'Proses Cuci' | 
    'Siap Diantar' | 
    'Selesai';

export interface LaundryOrder {
    id: string;
    customerId: string;
    customerName: string;
    customerAvatar: string;
    serviceType: LaundryServiceType | string;
    address: string;
    pickupTime: string;
    createdAt: string;
    estimatedPrice: number;
    statusSteps: LaundryStatus[];
    notes?: string;
    items?: LaundryOrderItem[];
    estimatedWeight?: number;
}